Add IGuideManager service
